cambiar disponibilidad

+++ scripts/create_lambda_zips.sh
@@ -3,51 +3,32 @@
 # Crear directorio temporal para los archivos
 mkdir -p temp_lambda
 
-# Copiar los archivos Python al directorio temporal
-cp lambda_vianda_buy.py temp_lambda/lambda_vianda_buy.py
-cd temp_lambda
-zip ../lambda_vianda_buy.zip lambda_vianda_buy.py
-cd ..
+# Lista de lambdas
+LAMBDAS=(
+  lambda_vianda_buy
+  lambda_rds_init
+  lambda_vianda_create
+  lambda_vianda_delete
+  lambda_vianda_list
+  lambda_vianda_get
+  lambda_vianda_profile
+  lambda_vianda_profile_update
+  lambda_cognito_post_confirmation
+  lambda_vianda_toggle_disponibilidad
+)
 
-cp lambda_rds_init.py temp_lambda/lambda_rds_init.py
-cd temp_lambda
-zip ../lambda_rds_init.zip lambda_rds_init.py
-cd ..
+for LAMBDA in "${LAMBDAS[@]}"; do
+  if [ -f scripts/${LAMBDA}.py ]; then
+    cp scripts/${LAMBDA}.py temp_lambda/${LAMBDA}.py
+    cd temp_lambda
+    zip ../scripts/${LAMBDA}.zip ${LAMBDA}.py
+    cd ..
+  else
+    echo "No se encontró scripts/${LAMBDA}.py, se omite."
+  fi
+  rm -f temp_lambda/${LAMBDA}.py
 
-cp lambda_vianda_create.py temp_lambda/lambda_vianda_create.py
-cd temp_lambda
-zip ../lambda_vianda_create.zip lambda_vianda_create.py
-cd ..
-
-cp lambda_vianda_delete.py temp_lambda/lambda_vianda_delete.py
-cd temp_lambda
-zip ../lambda_vianda_delete.zip lambda_vianda_delete.py
-cd ..
-
-cp lambda_vianda_list.py temp_lambda/lambda_vianda_list.py
-cd temp_lambda
-zip ../lambda_vianda_list.zip lambda_vianda_list.py
-cd ..
-
-cp lambda_vianda_get.py temp_lambda/lambda_vianda_get.py
-cd temp_lambda
-zip ../lambda_vianda_get.zip lambda_vianda_get.py
-cd ..
-
-cp lambda_vianda_profile.py temp_lambda/lambda_vianda_profile.py
-cd temp_lambda
-zip ../lambda_vianda_profile.zip lambda_vianda_profile.py
-cd ..
-
-cp lambda_vianda_profile_update.py temp_lambda/lambda_vianda_profile_update.py
-cd temp_lambda
-zip ../lambda_vianda_profile_update.zip lambda_vianda_profile_update.py
-cd ..
-
-cp lambda_cognito_post_confirmation.py temp_lambda/lambda_cognito_post_confirmation.py
-cd temp_lambda
-zip ../lambda_cognito_post_confirmation.zip lambda_cognito_post_confirmation.py
-cd ..
+done
 
 # Limpiar
-rm -rf temp_lambda 
+test -d temp_lambda && rm -rf temp_lambda
